+++ app/build.gradle
@@ -27,7 +27,7 @@ android {
         targetCompatibility JavaVersion.VERSION_1_8
     }
 
-    buildFeatures {
+    buildFeatures{
         viewBinding true
     }
 }
