+++ app/src/main/java/algonquin/cst2335/androidfinalproj/mainactivity/MainViewModel.java
@@ -1,4 +1,4 @@
-package algonquin.cst2335.androidfinalproj.currencyconverter.data;
+package algonquin.cst2335.androidfinalproj.mainactivity;
 
 import androidx.lifecycle.ViewModel;
