+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".currencyconverter.ui.MainActivity">
+    tools:context=".mainactivity.MainActivity">
 
     <TextView
         android:id="@+id/welcomeLabel"
